- merge LNET from Lustre 1.8    (o) rewritten socklnd and probably ptllnd    (o) its funny cause a lot of local changes we've made        have also been made    (o) their logging subsystem has greatly improved but        still fits in pretty easily with the mighty psclog    (o) deploy <inttypes.h> in some places since a deep        LNET dependency ended up doing it for us, which        was bad.    (o) fix a number of warnings of signed/unsigned        comparisions and unused variables in the new LNET        code
  some interface adjustments (subject to further change):

   old name		new name
   =======================================
    TCPLND_USESDP	USOCK_USESDP
    TCPLND_PORT		LNET_ACCEPT_PORT
    TCPLND_CPORT	USOCK_CPORT
    TCPLND_SERVER	<zapped>
    TCPLND_PORTINC	USOCK_PORTINC
    TCPLND_BUFSIZ	USOCK_BUFSIZ
    -DXT3CLIENT_NOSYS	-DCRAY_XT3

  i commented out the new pid/daemon code until we get the
  merger worked out.

  a number of these new LNET-related files can probably be
  pruned from the tree but i like to work in incremental
  steps so i don't lose progress.

  there may still be a number of changes we have made to the
  old socklnd over the years that must be checked to see if
  they've already been integrated into the new socklnd, or if
  we must redo these changes ourselves, again.

  don't upgrade your tree and expect everything to work the
  new LNET to work perfectly in every way as of yet...

  the new LNET apparently boasts support for apple/darwin
  and solaris (obviously), which will be nice for slash,
  but i missed merging a few select parts, so support for
  that will be even furtherly forthcoming.

- add code which detects environment variable settings of
  old LNET/LND settings and triggers warnings against usage.



git-svn-id: svn+ssh://frodo/cluster/svn/projects@5186 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ lnet-lite/ulnds/socklnd/handlers.c
@@ -377,7 +377,7 @@ usocklnd_activeconn_hellorecv(usock_conn_t *conn)
         peer->up_last_alive = cfs_time_current();
         
         /* peer says that we lost the race */
-        if (hello->kshm_ctype == SOCKLND_CONN_NONE) {
+        if (hello->kshm_ctype == (__u32)SOCKLND_CONN_NONE) {
                 /* Start new active conn, relink txs and zc_acks from
                  * the conn to new conn, schedule removing the conn.
                  * Actually, we're expecting that a passive conn will
@@ -941,7 +941,7 @@ usocklnd_send_tx(usock_conn_t *conn, usock_tx_t *tx)
                 do { 
                         LASSERT (tx->tx_niov > 0); 
                         
-                        if (nob < iov->iov_len) { 
+                        if ((size_t)nob < iov->iov_len) { 
                                 iov->iov_base = (void *)(((unsigned long)(iov->iov_base)) + nob);
                                 iov->iov_len -= nob; 
                                 break; 
@@ -994,7 +994,7 @@ usocklnd_read_data(usock_conn_t *conn)
                 do { 
                         LASSERT (conn->uc_rx_niov > 0); 
                         
-                        if (nob < iov->iov_len) { 
+                        if ((size_t)nob < iov->iov_len) { 
                                 iov->iov_base = (void *)(((unsigned long)(iov->iov_base)) + nob); 
                                 iov->iov_len -= nob; 
                                 break;
